+++ spring-cloud-starter-tencent-polaris-config/src/test/java/com/tencent/cloud/polaris/config/spring/annotation/SpringValueProcessorTest.java
@@ -22,7 +22,6 @@ import java.io.IOException;
 import java.lang.reflect.Method;
 import java.net.ServerSocket;
 import java.util.Collection;
-import java.util.Objects;
 import java.util.Optional;
 
 import com.tencent.cloud.polaris.config.PolarisConfigBootstrapAutoConfiguration;
@@ -31,10 +30,9 @@ import com.tencent.cloud.polaris.config.spring.property.Person;
 import com.tencent.cloud.polaris.config.spring.property.SpringValue;
 import com.tencent.cloud.polaris.config.spring.property.SpringValueRegistry;
 import com.tencent.polaris.api.utils.CollectionUtils;
-import org.junit.AfterClass;
-import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -47,6 +45,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.ImportResource;
 import org.springframework.stereotype.Component;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 /**
  * Test for {@link SpringValueProcessor}.
  *
@@ -56,8 +56,8 @@ public class SpringValueProcessorTest {
 
 	private static ServerSocket serverSocket;
 
-	@BeforeClass
-	public static void before() {
+	@BeforeAll
+	static void beforeAll() {
 		new Thread(() -> {
 			try {
 				serverSocket = new ServerSocket(8093);
@@ -69,8 +69,8 @@ public class SpringValueProcessorTest {
 		}).start();
 	}
 
-	@AfterClass
-	public static void after() throws IOException {
+	@AfterAll
+	static void afterAll() throws IOException {
 		serverSocket.close();
 	}
 
@@ -92,19 +92,19 @@ public class SpringValueProcessorTest {
 			PolarisConfigAutoConfiguration polarisConfigAutoConfiguration = context.getBean(PolarisConfigAutoConfiguration.class);
 			BeanFactory beanFactory = polarisConfigAutoConfiguration.beanFactory;
 			Collection<SpringValue> timeout = springValueRegistry.get(beanFactory, "timeout");
-			Assert.assertFalse(CollectionUtils.isEmpty(timeout));
+			assertThat(CollectionUtils.isEmpty(timeout)).isFalse();
 			Optional<SpringValue> springValueOptional = timeout.stream().findAny();
-			Assert.assertTrue(springValueOptional.isPresent());
+			assertThat(springValueOptional.isPresent()).isTrue();
 
 			SpringValue springValue = springValueOptional.get();
-			Assert.assertEquals("${timeout:1000}", springValue.getPlaceholder());
-			Assert.assertTrue(springValue.isField());
-			Assert.assertTrue(Objects.nonNull(springValue.getField()));
-			Assert.assertEquals("timeout", springValue.getField().getName());
-			Assert.assertEquals(int.class, springValue.getTargetType());
+			assertThat(springValue.getPlaceholder()).isEqualTo("${timeout:1000}");
+			assertThat(springValue.isField()).isTrue();
+			assertThat(springValue.getField()).isNotNull();
+			assertThat(springValue.getField().getName()).isEqualTo("timeout");
+			assertThat(springValue.getTargetType()).isEqualTo(int.class);
 
 			ValueTest bean = context.getBean(ValueTest.class);
-			Assert.assertEquals(10000, bean.timeout);
+			assertThat(bean.timeout).isEqualTo(10000);
 		});
 	}
 
@@ -127,19 +127,19 @@ public class SpringValueProcessorTest {
 			PolarisConfigAutoConfiguration polarisConfigAutoConfiguration = context.getBean(PolarisConfigAutoConfiguration.class);
 			BeanFactory beanFactory = polarisConfigAutoConfiguration.beanFactory;
 			Collection<SpringValue> name = springValueRegistry.get(beanFactory, "name");
-			Assert.assertFalse(CollectionUtils.isEmpty(name));
+			assertThat(name).isNotEmpty();
 			Optional<SpringValue> springValueOptional = name.stream().findAny();
-			Assert.assertTrue(springValueOptional.isPresent());
+			assertThat(springValueOptional.isPresent()).isTrue();
 
 			SpringValue springValue = springValueOptional.get();
 			Method method = springValue.getMethodParameter().getMethod();
-			Assert.assertTrue(Objects.nonNull(method));
-			Assert.assertEquals("setName", method.getName());
-			Assert.assertEquals("${name:1000}", springValue.getPlaceholder());
-			Assert.assertFalse(springValue.isField());
-			Assert.assertEquals(String.class, springValue.getTargetType());
+			assertThat(method).isNotNull();
+			assertThat(method.getName()).isEqualTo("setName");
+			assertThat(springValue.getPlaceholder()).isEqualTo("${name:1000}");
+			assertThat(springValue.isField()).isFalse();
+			assertThat(springValue.getTargetType()).isEqualTo(String.class);
 
-			Assert.assertEquals("test", ValueTest.name);
+			assertThat(ValueTest.name).isEqualTo("test");
 		});
 	}
 
@@ -163,31 +163,31 @@ public class SpringValueProcessorTest {
 			SpringValueRegistry springValueRegistry = context.getBean(SpringValueRegistry.class);
 			BeanFactory beanFactory = person.getBeanFactory();
 			Collection<SpringValue> name = springValueRegistry.get(beanFactory, "name");
-			Assert.assertFalse(CollectionUtils.isEmpty(name));
+			assertThat(name).isNotEmpty();
 			Optional<SpringValue> nameSpringValueOptional = name.stream().findAny();
-			Assert.assertTrue(nameSpringValueOptional.isPresent());
+			assertThat(nameSpringValueOptional.isPresent()).isTrue();
 
 			SpringValue nameSpringValue = nameSpringValueOptional.get();
 			Method method = nameSpringValue.getMethodParameter().getMethod();
-			Assert.assertTrue(Objects.nonNull(method));
-			Assert.assertEquals("setName", method.getName());
-			Assert.assertEquals("${name:test}", nameSpringValue.getPlaceholder());
-			Assert.assertFalse(nameSpringValue.isField());
-			Assert.assertEquals(String.class, nameSpringValue.getTargetType());
+			assertThat(method).isNotNull();
+			assertThat(method.getName()).isEqualTo("setName");
+			assertThat(nameSpringValue.getPlaceholder()).isEqualTo("${name:test}");
+			assertThat(nameSpringValue.isField()).isFalse();
+			assertThat(nameSpringValue.getTargetType()).isEqualTo(String.class);
 
 
 			Collection<SpringValue> age = springValueRegistry.get(beanFactory, "age");
-			Assert.assertFalse(CollectionUtils.isEmpty(age));
+			assertThat(age).isNotEmpty();
 			Optional<SpringValue> ageSpringValueOptional = age.stream().findAny();
-			Assert.assertTrue(ageSpringValueOptional.isPresent());
+			assertThat(ageSpringValueOptional.isPresent()).isTrue();
 
 			SpringValue ageSpringValue = ageSpringValueOptional.get();
 			Method method1 = ageSpringValue.getMethodParameter().getMethod();
-			Assert.assertTrue(Objects.nonNull(method1));
-			Assert.assertEquals("setAge", method1.getName());
-			Assert.assertEquals("${age:10}", ageSpringValue.getPlaceholder());
-			Assert.assertFalse(ageSpringValue.isField());
-			Assert.assertEquals(String.class, ageSpringValue.getTargetType());
+			assertThat(method1).isNotNull();
+			assertThat(method1.getName()).isEqualTo("setAge");
+			assertThat(ageSpringValue.getPlaceholder()).isEqualTo("${age:10}");
+			assertThat(ageSpringValue.isField()).isFalse();
+			assertThat(ageSpringValue.getTargetType()).isEqualTo(String.class);
 		});
 
 	}