+++ spring-cloud-starter-tencent-polaris-router/src/test/java/com/tencent/cloud/polaris/router/config/properties/PolarisNearByRouterPropertiesTest.java
@@ -18,12 +18,10 @@
 
 package com.tencent.cloud.polaris.router.config.properties;
 
-
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
 
 /**
  * test for {@link PolarisNearByRouterProperties}.
@@ -32,8 +30,8 @@ public class PolarisNearByRouterPropertiesTest {
 
 	PolarisNearByRouterProperties properties;
 
-	@Before
-	public void setUp() {
+	@BeforeEach
+	void setUp() {
 		properties = new PolarisNearByRouterProperties();
 	}